% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ctmle.R
\name{ctmle}
\alias{ctmle}
\title{ctmle}
\usage{
ctmle(logitQnA1, logitQnA0, W, A, Y, v = 10, gbounds = c(0, 1),
  searchstrategy = "ForwardStepwise", patience = 100,
  penalize_risk = TRUE, param = "ATE")
}
\arguments{
\item{logitQnA1}{numeric vector containing initial estimates for treatment group in logit scale.}

\item{logitQnA0}{numeric vector containing initial estimates for control group in logit scale.}

\item{W}{numeric matrix or dataframe, stand for covariates.}

\item{A}{numeric 1-0 vector, stand for treatment.}

\item{Y}{numeric vector, stand for outcome.}

\item{v}{integer, stand for number of folds for cross-validation in cTMLE.}

\item{searchstrategy}{Strategy for adding covariates to estimates of g. Default is "ForwardStepwise". Other options: "LogisticOrdering", "PartialCorrOrdering",  "SuperLearner"}

\item{patience}{Integer. For how many steps should CV continue after finding a local optimum? Defaults to typemax(Int).}

\item{param}{a string stand for different target parameter. "ATE" stands for additive  treatment effect, Mean1 and Mean0 stand for the mean outcome in treatment or control group.}

\item{gbounds:}{a  vector of numeric of length two (with both of them in [0,1]), stand for the truncation boud for gn.}
}
\value{
A list for targeted estimator and standard error.
}
\description{
A function compute ctmle estimator and its standard error based on julia package TargetedLearning.jl.
}
\author{
Cheng JU
}

