% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ctmle.R
\name{tmle}
\alias{tmle}
\title{tmle}
\usage{
tmle(logitQnA1, logitQnA0, W, A, Y, gn1, param = "ATE")
}
\arguments{
\item{logitQnA1}{logitQnA1 numeric vector containing initial estimates for treatment group in logit scale.}

\item{logitQnA0}{logitQnA0 numeric vector containing initial estimates for control group in logit scale.}

\item{W}{numeric matrix or dataframe, stand for covariates.}

\item{A}{numeric 1-0 vector, stand for treatment.}

\item{Y}{numeric vector, stand for outcome.}

\item{gn1}{numeric vector, stand for estimated probability of treatment (propensity score).}

\item{param}{a string stand for different target parameter. "ATE" stands for additive  treatment effect, Mean1 and Mean0 stand for the mean outcome in treatment or control group.}
}
\value{
A list for targeted estimator and standard error.
}
\description{
A function compute ctmle estimator and its standard error based on julia package TargetedLearning.jl.
}
\author{
Cheng JU
}

